DCC-213 Add authorization to all REST calls (as necessary)

+++ server/src/main/java/org/icgc/dcc/web/NextReleaseResource.java
@@ -12,9 +12,11 @@ import javax.ws.rs.core.Request;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
+import org.apache.shiro.SecurityUtils;
 import org.icgc.dcc.release.NextRelease;
 import org.icgc.dcc.release.ReleaseService;
 import org.icgc.dcc.release.model.Release;
+import org.icgc.dcc.shiro.AuthorizationPrivileges;
 
 import com.google.inject.Inject;
 
@@ -32,6 +34,9 @@ public class NextReleaseResource {
 
   @POST
   public Response release(String nextReleaseName, @Context Request req) {
+    if(SecurityUtils.getSubject().isPermitted(AuthorizationPrivileges.RELEASE_CLOSE.toString()) == false) {
+      return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+    }
     NextRelease oldRelease = releaseService.getNextRelease();
     // Check the timestamp of the oldRelease, since that is the object being updated
     ResponseTimestamper.evaluate(req, oldRelease.getRelease());
@@ -51,6 +56,11 @@ public class NextReleaseResource {
   @POST
   @Path("queue")
   public Response queue(List<String> projectKeys, @Context Request req) {
+    for(String projectKey : projectKeys) {
+      if(SecurityUtils.getSubject().isPermitted(AuthorizationPrivileges.projectViewPrivilege(projectKey)) == false) {
+        return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+      }
+    }
     ResponseTimestamper.evaluate(req, this.releaseService.getNextRelease().getRelease());
 
     if(this.releaseService.hasProjectKey(projectKeys)) {
@@ -64,6 +74,9 @@ public class NextReleaseResource {
   @DELETE
   @Path("queue")
   public Response removeAllQueued() {
+    if(SecurityUtils.getSubject().isPermitted(AuthorizationPrivileges.QUEUE_DELETE.toString()) == false) {
+      return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+    }
     this.releaseService.deleteQueuedRequest();
 
     return Response.ok().build();
@@ -80,6 +93,9 @@ public class NextReleaseResource {
   @POST
   @Path("signed")
   public Response signOff(List<String> projectKeys, @Context Request req) {
+    if(SecurityUtils.getSubject().isPermitted(AuthorizationPrivileges.RELEASE_SIGNOFF.toString()) == false) {
+      return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+    }
     ResponseTimestamper.evaluate(req, this.releaseService.getNextRelease().getRelease());
 
     if(this.releaseService.hasProjectKey(projectKeys)) {
@@ -93,6 +109,9 @@ public class NextReleaseResource {
   @PUT
   @Path("update")
   public Response update(Release release, @Context Request req) {
+    if(SecurityUtils.getSubject().isPermitted(AuthorizationPrivileges.RELEASE_MODIFY.toString()) == false) {
+      return Response.status(Status.UNAUTHORIZED).entity(new ServerErrorResponseMessage("Unauthorized")).build();
+    }
     if(release != null) {
       ResponseTimestamper.evaluate(req, release);
